refactor: move mesh file types into seperate src file
