+++ src/convertor/__init__.py
@@ -1,1 +1,2 @@
-from .convertor import convert_mesh, FileFormat, OutputMeshFile, OBJOutputMeshFile, STLOutputMeshFile
+from .convertor import convert_mesh
+from .mesh_types import FileFormat, OutputMeshFile, InputMeshFile, OBJOutputMeshFile, STLOutputMeshFile, OBJInputMeshFile, STLInputMeshFile
